Refactored existing models into JSON folder
