+++ samples/WebServerOAuthFlow/Controllers/HomeController.cs
@@ -3,6 +3,7 @@ using System.Web;
 using System.Web.Mvc;
 using Salesforce.Common;
 using Salesforce.Common.Models;
+using Salesforce.Common.Models.Json;
 
 namespace WebServerOAuthFlow.Controllers
 {
