+++ src/CommonLibrariesForNET.FunctionalTests/CommonTests.cs
@@ -5,6 +5,7 @@ using System.Net.Http;
 using NUnit.Framework;
 using Salesforce.Common.FunctionalTests.Models;
 using Salesforce.Common.Models;
+using Salesforce.Common.Models.Json;
 
 namespace Salesforce.Common.FunctionalTests
 {
