+++ src/CommonLibrariesForNET.UnitTests/AuthenticationClientRouteHandler.cs
@@ -4,6 +4,7 @@ using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 using Salesforce.Common.Models;
+using Salesforce.Common.Models.Json;
 
 namespace Salesforce.Common.UnitTests
 {
