+++ src/CommonLibrariesForNET/ForceAuthException.cs
@@ -1,4 +1,5 @@
 ﻿using Salesforce.Common.Models;
+using Salesforce.Common.Models.Json;
 
 namespace Salesforce.Common
 {
