+++ src/CommonLibrariesForNET/ForceException.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using Salesforce.Common.Models;
+using Salesforce.Common.Models.Json;
 
 namespace Salesforce.Common
 {
