+++ src/CommonLibrariesForNET/Models/Json/ErrorResponses.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace Salesforce.Common.Models
+namespace Salesforce.Common.Models.Json
 {
     public class ErrorResponses : List<ErrorResponse> { }
 }
