+++ src/CommonLibrariesForNET/Models/Json/SuccessResponse.cs
@@ -1,6 +1,6 @@
 ﻿using Newtonsoft.Json;
 
-namespace Salesforce.Common.Models
+namespace Salesforce.Common.Models.Json
 {
     public class SuccessResponse
     {
